+++ app/domain/work_prices/__init__.py
@@ -1,0 +1,11 @@
+from .entities import WorkPrice
+from .errors import WorkPriceNotFoundError, WorkPriceValidationError
+from .policies import ALLOWED_WORK_PRICE_CATEGORIES, validate_work_price_category
+
+__all__ = [
+    "ALLOWED_WORK_PRICE_CATEGORIES",
+    "WorkPrice",
+    "WorkPriceNotFoundError",
+    "WorkPriceValidationError",
+    "validate_work_price_category",
+]