+++ app/routes/__init__.py
@@ -24,7 +24,6 @@ from .namespaces.user_ns import user_ns
 from .namespaces.work_category_ns import work_category_ns
 from .namespaces.work_material_relation_ns import work_material_relation_ns
 from .namespaces.work_ns import work_ns
-from .namespaces.work_price_ns import work_price_ns
 
 
 def register_routes(app: Flask):
@@ -42,7 +41,6 @@ def register_namespaces(api: Api):
     api.add_namespace(object_ns)
     api.add_namespace(project_ns)
     api.add_namespace(work_ns)
-    api.add_namespace(work_price_ns)
     api.add_namespace(material_ns)
     api.add_namespace(work_material_relation_ns)
     api.add_namespace(project_material_ns)